refactor: DependencyRuleTest 부활
resolver 들 resolver 패키지로 이동
