+++ src/main/java/com/meoguri/linkocean/configuration/resolver/GetBookmarkQueryParams.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.controller.bookmark.support;
+package com.meoguri.linkocean.configuration.resolver;
 
 import java.util.List;
