+++ src/main/java/com/meoguri/linkocean/configuration/resolver/ProfileSearchTabConverterFactory.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.controller.profile.support;
+package com.meoguri.linkocean.configuration.resolver;
 
 import java.util.Arrays;
